+++ src/init.cpp
@@ -115,6 +115,9 @@
 #include <zmq/zmqrpc.h>
 #endif
 
+using common::AmountErrMsg;
+using common::InvalidPortErrMsg;
+using common::ResolveErrMsg;
 using kernel::DumpMempool;
 using kernel::LoadMempool;
 using kernel::ValidationCacheSizes;
