+++ src/node/mempool_args.cpp
@@ -8,19 +8,20 @@
 #include <kernel/mempool_options.h>
 
 #include <common/args.h>
+#include <common/messages.h>
 #include <consensus/amount.h>
 #include <kernel/chainparams.h>
 #include <logging.h>
 #include <policy/feerate.h>
 #include <policy/policy.h>
 #include <tinyformat.h>
-#include <util/error.h>
 #include <util/moneystr.h>
 #include <util/translation.h>
 
 #include <chrono>
 #include <memory>
 
+using common::AmountErrMsg;
 using kernel::MemPoolLimits;
 using kernel::MemPoolOptions;
