+++ src/rpc/util.cpp
@@ -7,6 +7,7 @@
 #include <clientversion.h>
 #include <core_io.h>
 #include <common/args.h>
+#include <common/messages.h>
 #include <common/types.h>
 #include <consensus/amount.h>
 #include <script/interpreter.h>
@@ -32,6 +33,8 @@
 #include <utility>
 
 using common::PSBTError;
+using common::PSBTErrorString;
+using common::TransactionErrorString;
 
 const std::string UNIX_EPOCH_TIME = "UNIX epoch time";
 const std::string EXAMPLE_ADDRESS[2] = {"bc1q09vm5lfy0j5reeulh4x5752q25uqqvz34hufdl", "bc1q02ad21edsxd23d32dfgqqsz4vv4nmtfzuklhy3"};
