+++ src/test/fuzz/kitchen_sink.cpp
@@ -2,19 +2,21 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <common/messages.h>
 #include <merkleblock.h>
 #include <policy/fees.h>
 #include <rpc/util.h>
 #include <test/fuzz/FuzzedDataProvider.h>
 #include <test/fuzz/fuzz.h>
 #include <test/fuzz/util.h>
-#include <util/error.h>
 #include <util/translation.h>
 
 #include <array>
 #include <cstdint>
 #include <optional>
 #include <vector>
+
+using common::TransactionErrorString;
 
 namespace {
 constexpr TransactionError ALL_TRANSACTION_ERROR[] = {
