+++ src/wallet/wallet.cpp
@@ -11,6 +11,7 @@
 #include <chain.h>
 #include <coins.h>
 #include <common/args.h>
+#include <common/messages.h>
 #include <common/settings.h>
 #include <common/signmessage.h>
 #include <common/system.h>
@@ -50,7 +51,6 @@
 #include <uint256.h>
 #include <univalue.h>
 #include <util/check.h>
-#include <util/error.h>
 #include <util/fs.h>
 #include <util/fs_helpers.h>
 #include <util/moneystr.h>
@@ -81,6 +81,8 @@
 
 struct KeyOriginInfo;
 
+using common::AmountErrMsg;
+using common::AmountHighWarn;
 using common::PSBTError;
 using interfaces::FoundBlock;
